refact: rename `mod serialization` -> `serde`; split out `json`, `ron`
