+++ src/lib.rs
@@ -27,7 +27,7 @@ pub use crate::cookie_path::CookiePath;
 mod cookie_store;
 pub use crate::cookie_store::{CookieStore, StoreAction};
 #[cfg(feature = "serde")]
-pub mod serialization;
+pub mod serde;
 mod utils;
 
 #[derive(Debug)]
